changed all datatypes in dataStructures (and other classes) to int & double removed custom resize functions removed fullScreen, minimize, move, exit buttons from all the screens set default stage style added listeners to native resize functions to update ui added separate scenes for every screen removed splash screen completely fixed sound icon and sound vBox location and usage for all screens added hover color effect for radioButtons, checkBoxes, comboBoxes, scrollpanes etc. completely removed separate ui views based on OS made all animations and transitions considerably faster fixed images in gridview in atlasScreen which were unable to show up converted all data files and output files to json and changed all read and write methods accordingly removed dependency of jdom library and added jackson libraries to process json files updated controlsfx library to snapshot version 12 moved some common ui recalculations to coreScreen converted all coordinates in attractions data file to correct format in order to open google maps correctly moved some counters for data files to their correspondent classes from GlobalVariables to dataStructures made a lot of changes and updates to ratioProperties class and data file updated name of FYROM to North Macedonia and corrected its position in countries data file removed all animations and sounds when minimizing, exiting or going full screen updated game properties screen to use ratio properties data removed all unused icons and images, used one globe stand image for both stands probably more stuff...

+++ src/main/java/com/johnzaro/geographyquiz/dataStructures/Area.java
@@ -4,12 +4,12 @@ import static com.johnzaro.geographyquiz.core.GlobalVariables.*;
 
 public class Area
 {
-	private float areaInKilometers;
-	private float areaInMiles;
+	private double areaInKilometers;
+	private double areaInMiles;
 
 	private int globalRanking;
 
-	private float percentOfWater;
+	private double percentOfWater;
 
 	private int coastlineInKilometers;
 	private int coastlineInMiles;
@@ -35,22 +35,22 @@ public class Area
 		else return numberFormatForUI.format(areaInMiles) + " sq mi";
 	}
 
-	public float getAreaInKilometers()
+	public double getAreaInKilometers()
 	{
 		return areaInKilometers;
 	}
 
-	public void setAreaInKilometers(float areaInKilometers)
+	public void setAreaInKilometers(double areaInKilometers)
 	{
 		this.areaInKilometers = areaInKilometers;
 	}
 
-	public float getAreaInMiles()
+	public double getAreaInMiles()
 	{
 		return areaInMiles;
 	}
 
-	public void setAreaInMiles(float areaInMiles)
+	public void setAreaInMiles(double areaInMiles)
 	{
 		this.areaInMiles = areaInMiles;
 	}
@@ -65,12 +65,12 @@ public class Area
 		this.globalRanking = globalRanking;
 	}
 
-	public float getPercentOfWater()
+	public double getPercentOfWater()
 	{
 		return percentOfWater;
 	}
 
-	public void setPercentOfWater(float percentOfWater)
+	public void setPercentOfWater(double percentOfWater)
 	{
 		this.percentOfWater = percentOfWater;
 	}
